COMP: Fix pow and specialization.
COMP(Win): Fix warning about conversion from size_t to uint.
Also add explicit parameter init = 0 in ComputeAllIndices,
instead of rely on default parameter.

COMP: Remove typename from non-template Monogenic...Test.

COMP: Solve unused variable in RieszFrequencyFunctionTest.

+++ Modules/Filtering/IsotropicWavelets/include/itkRieszRotationMatrix.hxx
@@ -267,7 +267,7 @@ RieszRotationMatrix<T, VImageDimension>::ComputeSteerableMatrix()
             }
             else
             {
-              rotationFactor *= std::pow(this->m_SpatialRotationMatrix[kIndice][dim], k);
+              rotationFactor *= std::pow(this->m_SpatialRotationMatrix[kIndice][dim], static_cast<int>(k));
             }
           }
         }
